move diagnostics utilities and completion_signatures into their own headers
